reogranized App; moved steps-components map to a module;
